Added file(s) for a a method to idea validate.
Inputs the data from the given csv files as problems and solutions and inputs that to the open ai API message list, and can thus give personalized requests about the data, if asked about what ideas are best and the ratings of each ( such as give me the 10 best ideas and the ratings of them ) It will give it. But it can be anything such as like what are the worst ideas, or whats the idea with the best scalability, etc.